requirements.txt
google-genai
python-dotenv
google-cloud-aiplatform[adk]
plotly
pandas
kaleido
